+++ packages/orq-rc/src/funcs/identitiesRetrieve.ts
@@ -3,7 +3,7 @@
  */
 
 import { OrqCore } from "../core.js";
-import { encodeSimple } from "../lib/encodings.js";
+import { encodeFormQuery, encodeSimple } from "../lib/encodings.js";
 import * as M from "../lib/matchers.js";
 import { compactMap } from "../lib/primitives.js";
 import { safeParse } from "../lib/schemas.js";
@@ -39,6 +39,7 @@ export function identitiesRetrieve(
   Result<
     operations.RetrieveIdentityResponseBody,
     | errors.RetrieveIdentityResponseBody
+    | errors.RetrieveIdentityIdentitiesResponseBody
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -65,6 +66,7 @@ async function $do(
     Result<
       operations.RetrieveIdentityResponseBody,
       | errors.RetrieveIdentityResponseBody
+      | errors.RetrieveIdentityIdentitiesResponseBody
       | OrqError
       | ResponseValidationError
       | ConnectionError
@@ -96,6 +98,10 @@ async function $do(
   };
   const path = pathToFunc("/v2/identities/{id}")(pathParams);
 
+  const query = encodeFormQuery({
+    "include_metrics": payload.include_metrics,
+  });
+
   const headers = new Headers(compactMap({
     Accept: "application/json",
   }));
@@ -125,6 +131,7 @@ async function $do(
     baseURL: options?.serverURL,
     path: path,
     headers: headers,
+    query: query,
     body: body,
     userAgent: client._options.userAgent,
     timeoutMs: options?.timeoutMs || client._options.timeoutMs || 600000,
@@ -136,7 +143,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["404", "4XX", "5XX"],
+    errorCodes: ["404", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -152,6 +159,7 @@ async function $do(
   const [result] = await M.match<
     operations.RetrieveIdentityResponseBody,
     | errors.RetrieveIdentityResponseBody
+    | errors.RetrieveIdentityIdentitiesResponseBody
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -163,6 +171,7 @@ async function $do(
   >(
     M.json(200, operations.RetrieveIdentityResponseBody$inboundSchema),
     M.jsonErr(404, errors.RetrieveIdentityResponseBody$inboundSchema),
+    M.jsonErr(500, errors.RetrieveIdentityIdentitiesResponseBody$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });