+++ packages/orq-rc/src/models/operations/createidentity.ts
@@ -139,7 +139,7 @@ export const CreateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-26T10:52:27.533Z",
+    "2026-03-27T06:06:37.886Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {